requirements.txt
appnope==0.1.0
decorator==4.1.2
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.6
django-extensions==1.9.8
django-mathfilters==0.4.0
django-publications==0.6.2
django-registration-redux==1.8
django-toolbelt==0.0.1
gunicorn==19.7.1
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
latexcodec==1.0.5
model-mommy==1.4.0
mpmath==1.0.0
olefile==0.44
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
Pygments==2.2.0
pylibmc==1.5.2
pytz==2017.2
PyYAML==3.12
simplegeneric==0.8.1
six==1.11.0
static3==0.7.0
sympy==1.1.1
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7
whitenoise==3.3.1
